+++ src/app/view/mission/new/NewMissionDatasourceOsmoseComponent.js
@@ -6,6 +6,43 @@ import Select from 'material-ui/Select';
 import Typography from 'material-ui/Typography';
 import Wait from '../../WaitComponent';
 
+/*
+ * Definition of editors according to error type
+ */
+const ERROR_TO_EDITORS = {
+	"1110": "all",
+	"1140": "all",
+	"1210": "all",
+	"2010": "all",
+	"2030": "all",
+	"2060": "all",
+	"2080": "all",
+	"2090": "all",
+	"2100": "all",
+	"2110": "all",
+	"2120": "all",
+	"2130": "all",
+	"2140": "all",
+	"3080": "all",
+	"3160": "all",
+	"3210": "all",
+	"3220": "all",
+	"3230": "all",
+	"3240": "all",
+	"4030": [ "disabled" ],
+	"4070": "all",
+	"7011": "all",
+	"7012": "all",
+	"7040": "all",
+	"7130": "all",
+	"7140": "all",
+	"7150": "all",
+	"7170": "all",
+	"7190": "all",
+	"8__1": "all",
+	"8__0": [ "disabled" ] //importer
+};
+
 /**
  * New mission datasource osmose component allows user to input settings for Osmose datasource
  */
@@ -15,7 +52,8 @@ class NewMissionDatasourceOsmoseComponent extends Component {
 		
 		this.state = {
 			items: null,
-			selectedItem: ""
+			selectedItem: "",
+			allowedEditors: "all"
 		};
 		
 		this.request = new OsmoseRequest();
@@ -40,8 +78,27 @@ class NewMissionDatasourceOsmoseComponent extends Component {
 	_changed(what, value) {
 		if(what === "item" && value !== this.state.selectedItem) {
 			if(value === "") { value = null; }
-			this.props.onChange({ item: value });
+			this.props.onChange({ item: value, allowedEditors: this._getEditorsForError({ id: value }) || this.state.allowedEditors });
 			this.setState({ selectedItem: value });
+		}
+	}
+	
+	_getEditorsForError(item) {
+		const itemId = item.id.toString();
+		if(ERROR_TO_EDITORS[itemId]) {
+			return ERROR_TO_EDITORS[itemId];
+		}
+		else {
+			const keyMatchers = Object.keys(ERROR_TO_EDITORS).filter(k => k.indexOf("_") >= 0);
+			for(const km of keyMatchers) {
+				if(
+					itemId.length == km.length
+					&& itemId.match(km.replace(/_/g, "[0-9]"))
+				) {
+					return ERROR_TO_EDITORS[km];
+				}
+			}
+			return null;
 		}
 	}
 	
@@ -84,7 +141,14 @@ class NewMissionDatasourceOsmoseComponent extends Component {
 			this.request
 			.fetchItems()
 			.then(items => {
-				items = items.filter(i => i.name && i.id && i.name.en);
+				items = items.filter(i => {
+					if(i.name && i.id && i.name.en) {
+						return this._getEditorsForError(i) !== null;
+					}
+					else {
+						return false;
+					}
+				});
 				this.setState({ items: items });
 			})
 			.catch(e => {